requirements.txt
tensorflow
numpy
matplotlib
seaborn
scikit-learn
fastapi
uvicorn
pillow
python-multipart
